pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf-toolbox-qt"
version = "0.1.0"
description = "Local offline PDF toolbox built with PySide6"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "PDF Toolbox Qt" }]

dependencies = [
  "PySide6>=6.5",
  "pikepdf>=8.0",
  "PyMuPDF>=1.23",
  "Pillow>=10.0",
  "pywin32>=306"
]

[project.scripts]
pdf-toolbox-qt = "pdf_toolbox.app:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
